requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
Flask-Migrate==4.0.4
Flask-WTF==1.0.1